Use long instead of int for test comparison

+++ tests/test_columnfamily.py
@@ -50,7 +50,7 @@ class TestColumnFamily(unittest.TestCase):
         columns = {'1': 'val1', '2': 'val2'}
         assert_raises(NotFoundException, cf.get, key)
         ts = cf.insert(key, columns)
-        assert_true(isinstance(ts, int))
+        assert_true(isinstance(ts, long))
         assert_equal(cf.get(key), columns)
 
     def test_insert_multiget(self):
